+++ src/verifiers/dnsText/openAttestationDnsTxt.ts
@@ -1,13 +1,8 @@
 import { getData, v2, v3, WrappedDocument } from "@govtechsg/open-attestation";
 import { getDocumentStoreRecords } from "@govtechsg/dnsprove";
 import { utils } from "ethers";
-import {
-  isWrappedV2Document,
-  OpenAttestationDnsTxtCode,
-  VerificationFragmentType,
-  VerificationManagerOptions,
-  Verifier
-} from "../types/core";
+import { isWrappedV2Document, VerificationFragmentType, VerificationManagerOptions, Verifier } from "../../types/core";
+import { OpenAttestationDnsTxtCode } from "../../types/error";
 
 interface Identity {
   status: "VALID" | "INVALID" | "SKIPPED";
