+++ src/verifiers/hash/openAttestationHash.ts
@@ -1,5 +1,6 @@
 import { verifySignature } from "@govtechsg/open-attestation";
-import { OpenAttestationHashCode, VerificationFragmentType, Verifier } from "../types/core";
+import { VerificationFragmentType, Verifier } from "../../types/core";
+import { OpenAttestationHashCode } from "../../types/error";
 
 const name = "OpenAttestationHash";
 const type: VerificationFragmentType = "DOCUMENT_INTEGRITY";
